+++ src/ModularInteger.php
@@ -17,17 +17,13 @@ class ModularInteger
 {
     /**
      * The value of this number.
-     *
-     * @var integer
      */
     public protected(set) int $value;
 
     /**
      * The modulus of the modular set of which this number is a part.
-     * 
+     *
      * For example, for the hours on a clock, the modulus is 12.
-     * 
-     * @var integer
      */
     public protected(set) int $modulus;
 
@@ -38,10 +34,9 @@ class ModularInteger
      * @param integer $modulus The modulus of the modular set of which this 
      * number is a part.
      * 
-     * Keep in mind that PHP always treats the right operand of the modulo 
-     * operation as an absolute (positive) value, meaning that every negative
-     * value of this parameter is treated as a positive value. The same is not
-     * true for a traditional calculator.
+     * Keep in mind that PHP always treats the modulus as absolute value, 
+     * meaning that every negative modulus is treated as a positive value. 
+     * The same is not true for a traditional calculator app.
      * @throws DivisionByZeroError when $modulus equals zero.
      */
     public function __construct(int $value, int $modulus)
@@ -53,25 +48,15 @@ class ModularInteger
     /**
      * Return true if this instance is congruent to $number,
      * false otherwise.
-     *
-     * @param ModularInteger $number
-     * @return boolean
-     * @throws DifferentModulusError when this instance has a modulus different
-     * from the modulus of $number. 
      */
     public function isCongruent(ModularInteger $number): bool
     {
-        if ($this->modulus != $number->modulus) throw new DifferentModulusError($this, $number);
+        if ($this->modulus != $number->modulus) return false;
         return ($this->value - $number->value) % $this->modulus == 0;
     }
 
     /**
      * Alias of isCongruent() method.
-     *
-     * @param ModularInteger $number
-     * @return boolean
-     * @throws DifferentModulusError when this instance has a modulus different
-     * from the modulus of $number. 
      */
     public function equals(ModularInteger $number): bool
     {
@@ -81,9 +66,7 @@ class ModularInteger
     /**
      * Sum this instance with $number.
      *
-     * @param ModularInteger $number
-     * @return ModularInteger
-     * @throws DifferentModulusError when this instance and $number have 
+     * @throws DifferentModulusError when this instance and $number have
      * different modulus.
      */
     public function add(ModularInteger $number): ModularInteger
@@ -93,9 +76,9 @@ class ModularInteger
 
     /**
      * Multiply this instance with $number.
-     *
-     * @param ModularInteger $number
-     * @return ModularInteger
+     * 
+     * @throws DifferentModulusError when this instance and $number have
+     * different modulus.
      */
     public function multiply(ModularInteger $number): ModularInteger
     {
@@ -104,9 +87,6 @@ class ModularInteger
 
     /**
      * Raise this instance to power $k.
-     *
-     * @param integer $k
-     * @return ModularInteger
      */
     public function power(int $k): ModularInteger
     {