+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## v2.0.0 - 2025-01-08
+### Changed
+- BREAKING CHANGE: Parameter name `$a` to `$base` in `IntegerModularExponentiation::__construct()` method.
+- API documentation.
+- Improvements to exponentiation operation.
+### Fixed
+- Negative base exponentiation causing not to throws `IntegerOverflowException` when needed.
+
 ## v1.0.0 - 2025-01-04
 ### Added
 - The `ModularInteger` class that represent an integer variable along with its modulus.